Reducing the complexity of the examples
Reducing the complexity of the examples, centralizing dependences in a singleton. So common it's all in one place and not have to follow references. It is not a major change in what, I know, but I think it simplifies the understanding of the common parts. For example, this makes it clearer than in the app are only two Scheduler (mainScheduler and backgroundWorkScheduler)