Reflect infinite sequences in the regex example
The text gives an example of an infinite sequence here "Some sequences are finite, and some are infinite, like sequence of button taps" and just few lines below tries to explain sequences with a regex (which is great). However the regex was always having a terminating element (error or completion) and that contradicts the idea of an infinite sequence (which never terminates). 

What I did was to add a "?" to the terminating element in the regex example