Add ForwardIf operator
ForwardIf propagates the source observable sequence while the condition
observable sequence last value is true.

This is useful when an Observable performs side effects on subscription, so we
only want to subscribe when another observable is true.

An example case for this would be an Observable that performs network requests.
We only want it to hit the network if the subscribing view controller is
visible.