Adds `isEnabled` property for UILabel extension
Despite the fact that `UILabel` is not an `UIControl`, it has an `isEnabled` property. Just out of curiosity, according to the docs (https://developer.apple.com/reference/uikit/uilabel/1620530-isenabled), "This property determines only how the label is drawn. Disabled text is dimmed somewhat to indicate it is not active. This property is set to true by default".