Use Signals for tapping events instead of Drivers
Driver does not fit to some UI events such as tapping.
Because we do not expect to get past "tap" events.

But Driver is using `share(replay: 1)`. So it may
replay accidentally the past event for some conditions.
This replay is useful only for observables like
Variables such as `rx.text`, but `rx.tap` is not
Variable like.

BTW, Signal can fit the usecase. Signal is using
`.share()`, so it does not replay past tapping events.